fix: discard trailing path information
Discord spoilers add `||` on either side of the content to put in a
spoiler, and when making an xfuraffinity link a spoiler, the `|` would
be added to the link, causing an invalid link error.

Also included in this commit is a refactor to extract data from
submissions more easily, and to prepare the support of oembeds for
Discord.